# docs/swagger.yaml
openapi: 3.0.2
info:
    title: Games API
    version: 1.0.0
    description: "### This is a Swagger UI for our Games API\r\n\r\n### Error handling\r\nErrors are presented using the HTTP status codes 400–599.\r\nThe response includes a more detailed explanation about the error in the body.\r\n\r\n\r\n### Generic API errors\r\nThese error responses can be returned by any of the described endpoints.\r\n\r\n| HTTP status | Description |\r\n|--------------|-------------|\r\n| 400 | Bad input parameters or invalid request. More information can be found in the response body. |\r\n| 401 | The consumer does not have access to the requested resource or the provided token is not valid. |\r\n| 404 | The requested resource cannot be found or the user does not have access to the resource. |\r\n| 5XX | A server error has occurred while processing the request. |\r\n"
paths:
    /games:
        get:
            tags:
                - games
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                List of games:
                                    value:
                                        - Witcher 3
                                        - Cyberpunk 2077
                    description: application/json
            summary: Get a list of all games
            description: Returns a list of all games in API
        post:
            requestBody:
                description: 'A new game to be created '
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/error'
                        examples:
                            Error:
                                value:
                                    error: 'Unable to add new game: invalid data format'
                required: true
            tags:
                - new-games
            responses:
                '201':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/new-game'
                    description: Add a game
                '400':
                    description: One or all params are missing
            summary: Creates a game
            description: Creates a new game
    '/games/{id}':
        get:
            tags:
                - details
            parameters:
                -
                    name: id
                    in: path
                    required: true
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                Full details of a given game:
                                    value:
                                        id: 1
                                        name: Witcher 3
                                        price: 29.99
                    description: Returns full details of a given id game
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                Game \w+ was not found:
                                    value:
                                        error: Game \w+ was not found in game array
                    description: Game was not found in games array
            operationId: id
            summary: Get full details of a game
            description: Returns full details of a given id game
        put:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/game-update'
                        examples:
                            Update a game:
                                value:
                                    name: Super Mario
                                    price: 59.99
            tags:
                - update
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/game-update'
                            examples:
                                The game has been updated:
                                    value:
                                        _id: 64f1234567890abcde123456
                                        name: Super Mario
                                        price: 59.99
                    description: Game updated successfully
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                Game not found:
                                    value:
                                        message: Game not found
                    description: Game not found
            summary: Update a game by ID
            description: 'Updates the details of a game identified by its ID. '
        delete:
            tags:
                - delete
            responses:
                '204':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/delete-game'
                    description: Game successfully deleted
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                    description: Game not found
            summary: Delete a game by its ID
            description: Deletes a game by ID. Returns an error if the game is not found.
        parameters:
            -
                name: id
                in: path
                required: true
components:
    schemas:
        error:
            title: Root Type for error
            description: ''
            type: object
            properties:
                error:
                    type: string
            example:
                error: Error message.
        new-game:
            title: Root Type for new-game
            description: ''
            required:
                - price
                - name
            type: object
            properties:
                name:
                    type: string
                price:
                    format: float
                    type: number
            example:
                name: Tetris
                price: 0.99
        delete-game:
            description: ''
            required:
                - id
            type: object
            properties:
                id:
                    description: ''
                    type: string
            example:
                id: 1
        game-update:
            title: Root Type for game-update
            description: ''
            type: object
            properties:
                name:
                    type: string
                price:
                    format: float
                    type: number
            example:
                name: Super Mario
                price: 59.99